requirements.txt
asgiref==3.5.0
autopep8==2.0.2
blinker==1.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==3.2.12
djangorestframework==3.14.0
filelock==3.11.0
gunicorn==20.1.0
httplib2==0.20.2
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
mypy==1.2.0
mypy-extensions==1.0.0
netifaces==0.11.0
oauthlib==3.2.0
Pillow==9.5.0
platformdirs==3.2.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
pytz==2022.1
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.2
systemd-python==234
tomli==2.0.1
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
virtualenv==20.21.0
wadllib==1.3.6
zipp==1.0.0
